+++ R/set_smc_options.R
@@ -6,6 +6,12 @@
 #' @param n_particles Integer specifying the number of particles.
 #' @param mcmc_steps Number of MCMC steps to be applied in the resample-move
 #'   step.
+#' @param resampler Character string defining the resampling method to use. One
+#'   of "stratified", "systematic", "residual", and "multinomial". Defaults to
+#'   "stratified". While multinomial resampling was used in
+#'   \insertCite{steinSequentialInferenceMallows2023;textual}{BayesMallows},
+#'   stratified, systematic, or residual resampling typically give lower Monte
+#'   Carlo error \insertCite{Douc2005,Hol2006,Naesseth2019}{BayesMallows}.
 #' @param latent_sampling_lag Parameter specifying the number of timesteps to go
 #'   back when resampling the latent ranks in the move step. See Section 6.2.3
 #'   of \insertCite{Kantas2015}{BayesMallows} for details. The \eqn{L} in their
@@ -14,6 +20,7 @@
 #'   are resampled. If set to `0`, no move step is applied to the latent ranks.
 #'
 #' @return An object of class "SMCOptions".
+#'
 #'
 #' @section Lag parameter in move step:
 #'
@@ -41,8 +48,8 @@
 #'   This means that with \eqn{L=0} no move step is performed on any latent
 #'   ranks, whereas \eqn{L=1} means that the move step is only applied to the
 #'   parameters entering at the given timestep. The default,
-#'   `latent_sampling_lag = NA` means that \eqn{L=t} at each timestep, and
-#'   hence all latent ranks are part of the move step at each timestep.
+#'   `latent_sampling_lag = NA` means that \eqn{L=t} at each timestep, and hence
+#'   all latent ranks are part of the move step at each timestep.
 #'
 #'
 #' @export
@@ -52,10 +59,15 @@
 set_smc_options <- function(
     n_particles = 1000,
     mcmc_steps = 5,
+    resampler = c("stratified", "systematic", "residual", "multinomial"),
     latent_sampling_lag = NA_integer_) {
   validate_integer(n_particles)
   validate_integer(mcmc_steps)
   if (!is.na(latent_sampling_lag)) validate_integer(latent_sampling_lag)
+  resampler <- match.arg(
+    resampler,
+    c("stratified", "systematic", "residual", "multinomial")
+  )
 
   ret <- as.list(environment())
   class(ret) <- "SMCOptions"